@@ -3,11 +3,11 @@ using System.Collections.Generic;
 
 namespace passportapi.Domain.Models
 {
-    public partial class PassportInfo
+    public partial class PassportIndex
     {
         public int Id { get; set; }
         public string Passport { get; set; }
         public string Destination { get; set; }
-        public Enumerations Value { get; set; }
+        public int Code { get; set; }
     }
 }
