@@ -8,15 +8,15 @@ namespace passportapi.Domain.Services.Communication
 {
     public class PassportInfoResponse: BaseResponse
     {
-        public PassportInfo PassportInfo { get; private set; }
+        public PassportIndex PassportIndex { get; private set; }
 
-        private PassportInfoResponse(bool success, string message, PassportInfo passportInfo): base (success, message)
+        private PassportInfoResponse(bool success, string message, PassportIndex passportIndex): base (success, message)
         {
-            PassportInfo = passportInfo;
+            PassportIndex = passportIndex;
         }
 
         //produces success message
-        public PassportInfoResponse(PassportInfo passportinfo) : this(true, string.Empty, passportinfo) { }
+        public PassportInfoResponse(PassportIndex passportinfo) : this(true, string.Empty, passportinfo) { }
 
         //produces error message
         public PassportInfoResponse(string message) : this(false, message, null) { }
